chore(web/playground): post-review polish — rename hook file, reorder classes, tighten CLAUDE.md
Address the three Minor findings from the final issue-32 code review:

- M-2: rename ChatModeTabs.tsx → useChatModeTabs.ts. The file's
  sole exports are the useChatModeTabs hook and the ChatModeTab
  type — there is no ChatModeTabs component anymore (the shim was
  removed in 395169e). Renaming aligns with the project's
  "final-form, no compat shim" policy. Three import sites updated.
- M-3: regroup the inline-image className in MessageList.tsx so
  alignment / sizing / object-fit / border read in canonical
  Tailwind order (self-start → max-h-64 w-auto max-w-full →
  rounded border border-border → object-contain). Pure cosmetic;
  test assertions use toContain so order is irrelevant for tests.
- M-4: replace "Optional sub-toolbar" in CLAUDE.md with the
  explicit OR rule (rendered when paramsSlot is non-null OR any
  of tabs / historySlot / viewCodeSnippets / toolbarRightSlot is
  non-empty), matching the actual showToolbar / showCollapseButton
  derivation in PlaygroundShell.tsx.

Issue: #32
