+++ apps/web/src/features/playground/chat/ChatPage.tsx
@@ -11,7 +11,7 @@ import { useTranslation } from "react-i18next";
 import { toast } from "sonner";
 import { CategoryEndpointSelector } from "../CategoryEndpointSelector";
 import { PlaygroundShell } from "../PlaygroundShell";
-import { useChatModeTabs } from "../chat-compare/ChatModeTabs";
+import { useChatModeTabs } from "../chat-compare/useChatModeTabs";
 import { genChatSnippets } from "../code-snippets/chat";
 import { HistoryDrawer } from "../history/HistoryDrawer";
 import {
